Exercise 2.10 Completed
