+++ part2/phonebook/src/components/Person.jsx
@@ -1,0 +1,3 @@
+const Person = (props) => <p>{props.name} {props.number}</p>
+
+export default Person
